# apps/api/swagger.yaml
openapi: 3.0.3
info:
  title: Brooklyn Fast Food Restaurant — API
  description: |
    REST API for the Brooklyn Fast Food Restaurant platform.

    ## Authentication
    Most endpoints require a Bearer token in the `Authorization` header.
    Obtain tokens via the `/auth/login` or `/auth/register` endpoints.

    ## Role Hierarchy
    | Role         | ID | Description                          |
    |-------------|----|--------------------------------------|
    | super-admin | 1  | Full system access                   |
    | admin       | 2  | User management (read/update)        |
    | clients     | 3  | Regular customers (default on signup)|
  version: 0.1.0
  contact:
    name: Brooklyn Fast Food Team
servers:
  - url: http://localhost:3001/api/v1
    description: Local development (v1)
  - url: http://localhost:3001/api
    description: Local development (legacy)
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using a refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  description: Valid refresh token
      responses:
        "200":
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token obtained from `/auth/login` or `/auth/register`.
        Send as `Authorization: Bearer <token>`.

  # ── Reusable Parameters ───────────────────────────────────
...
  responses:
...
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            message: No token provided
...
  schemas:
...
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
          description: Short-lived JWT (default 15 min)
        refreshToken:
          type: string
          description: Long-lived JWT (default 7 days)
...
    TokenResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/AuthTokens"
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string